JC - Added Programmalijst
- Ophalen data via API #47
- Variable gemaakt voor aanpassingen van dagen voor #24
- Voevoeging radio programma lijst #23
- Toegvoeging fields om alleen belangrijke fields te laten zien via 1 api request
- Added 2 componenten voor schedule en programmacard #20
- Animatie voor tekst als deze te lang is
- Sortering API data op from data
- Sortering radio stations met shows
- Programma card verkleining als data niet goed past

Pull-request voor #49
